requirements.txt
appdirs==1.4.4
APScheduler==3.6.3
certifi==2020.6.20
cffi==1.14.4
chardet==3.0.4
conda==4.3.16
cryptography==3.3.1
decorator==4.4.2
distlib==0.3.1
filelock==3.0.12
html2text==2020.1.16
idna==2.10
pipenv==2020.8.13
pycosat==0.6.3
pycparser==2.20
python-telegram-bot==13.1
pytz==2020.5
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
tornado==6.1
tzlocal==2.1
urllib3==1.25.10
virtualenv==20.0.31
virtualenv-clone==0.5.4
